# api-docs.yaml
openapi: 3.0.1
info:
  title: Instant Messaging API
  description: |
    API REST pour une messagerie instantanée qui comprend la gestion  des utilisateur·rice·s, des conversations et des messages. Cette API est utilisée dans  le cadre des modules 194, 195, 322, et 426 à l'EPAI Fribourg.
  contact:
    name: Jérôme Frossard
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: http://localhost:8080/api/v1
  description: Generated server url
security:
- BearerAuth: []
...
paths:
...
  /conversation/{conversationId}/participants/{participantId}:
    patch:
      tags:
      - participant
      summary: Modifie le rôle ou le statut d'un·e participant·e.
      description: |
        Modifie le rôle ou le statut d'un·e participant·e à partir de son identifiant.

        * Si la conversation est une conversation de groupe, le ou la propriétaire de la conversation peut bloquer un·e participant·e non propriétaire ou lui attribuer le rôle de propriétaire.

        * Si la conversation est une conversation privée, un·e participant·e peut changer son propre status d'actif à inactif. Cette opération est irréversible.
      operationId: update-participant
      parameters:
      - name: conversationId
        in: path
        description: Identifiant de la conversation
        required: true
        schema:
          type: integer
          format: int64
        example: 100
      - name: participantId
        in: path
        description: Identifiant de la conversation
        required: true
        schema:
          type: integer
          format: int64
        example: 11
      requestBody:
        description: Utilisateur·rice à créer
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParticipantUpdateDto"
            examples:
              Bloquer un utilisateur:
                description: Bloquer un utilisateur
                value:
                  status: BLOCKED
              Modifier le nom d'utilisateur·rice:
                description: Modifier le nom d'utilisateur·rice
                value:
                  username: "+41799999123"
              Modifier le nom affichable:
                description: Modifier le nom affichable
                value:
                  displayName: Sheana
        required: true
      responses:
        "429":
          description: Too Many Requests
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "200":
          description: Rôle ou statut de lea participant·modifié avec succès.
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/ParticipantResponseDto"
              example:
                userId: 15
                username: fallom@example.com
                role: MEMBER
                status: INACTIVE
                _links:
                  self:
                    href: http://localhost:8080/api/v1/conversation/110/participants/15
        "415":
          description: Unsupported Media Type
        "406":
          description: Not Acceptable
        "500":
          description: Internal Server Error
...
  /conversation/{conversationId}/participants:
    get:
      tags:
      - participant
      summary: Récupèrer la liste des participant·e·s d'une conversation.
      description: Récupère la liste des participant·e·s d'une conversation à partir
        de son identifiant.
      operationId: get-participants
      parameters:
      - name: conversationId
        in: path
        description: Identifiant de la conversation
        required: true
        schema:
          type: integer
          format: int64
        example: 110
      responses:
        "429":
          description: Too Many Requests
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "200":
          description: Liste de participants récupérée avec succès.
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/ParticipantResponseDto"
              example:
                _embedded:
                  participants:
                  - userId: 11
                    username: sheana@example.com
                    role: MEMBER
                    status: ACTIVE
                    _links:
                      self:
                        href: http://localhost:8080/api/v1/conversation/100/participants/11
                  - userId: 12
                    username: idaho@example.com
                    role: MEMBER
                    status: ACTIVE
                    _links:
                      self:
                        href: http://localhost:8080/api/v1/conversation/100/participants/12
                _links:
                  self:
                    href: http://localhost:8080/api/v1/conversation/100/participants
        "415":
          description: Unsupported Media Type
        "406":
          description: Not Acceptable
        "500":
          description: Internal Server Error
...
components:
  schemas:
...
    Links:
      type: object
      additionalProperties:
        $ref: "#/components/schemas/Link"
...
    ParticipantResponseDto:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        username:
          type: string
        role:
          type: string
        status:
          type: string
        _links:
          $ref: "#/components/schemas/Links"
...
    ParticipantUpdateDto:
      type: object
      properties:
        role:
          type: string
          enum:
          - OWNER
          - MEMBER
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
          - BLOCKED
...
    Link:
      type: object
      properties:
        href:
          type: string
        hreflang:
          type: string
        title:
          type: string
        type:
          type: string
        deprecation:
          type: string
        profile:
          type: string
        name:
          type: string
        templated:
          type: boolean
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer